// packages/datadog-api-client-v1/models/LogsByRetention.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogsByRetentionMonthlyUsage } from './LogsByRetentionMonthlyUsage';
import { LogsByRetentionOrgs } from './LogsByRetentionOrgs';
import { LogsRetentionAggSumUsage } from './LogsRetentionAggSumUsage';
/**
* Object containing logs usage data broken down by retention period.
*/
export declare class LogsByRetention {
    'orgs'?: LogsByRetentionOrgs;
    /**
    * Aggregated index logs usage for each retention period with usage.
    */
    'usage'?: Array<LogsRetentionAggSumUsage>;
    'usageByMonth'?: LogsByRetentionMonthlyUsage;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsByRetention;
    static serialize(data: LogsByRetention): {
        [key: string]: any;
    };
    constructor();
}